+++ src/HealthChecks.Azure.Messaging.EventHubs/AzureEventHubHealthCheck.cs
