+++ src/HealthChecks.Azure.Messaging.EventHubs/Configuration/AzureEventHubHealthCheckOptions.cs
